Refactored math package into seperate submodules
