+++ model/sphfluid.go
@@ -1,10 +1,11 @@
 package model
 
-import "github.com/andewx/dieselfluid/model/sph"
-import "github.com/andewx/dieselfluid/model/field"
-import V "github.com/andewx/dieselfluid/math/mgl"
-import G "github.com/andewx/dieselfluid/geom/mesh"
-import "github.com/andewx/dieselfluid/sphmethod"
+import (
+	"github.com/andewx/dieselfluid/model/sph"
+
+	G "github.com/andewx/dieselfluid/geom/mesh"
+	"github.com/andewx/dieselfluid/sphmethod"
+)
 
 //SPHFluid - Highest Level Fluid Model Abstraction
 type SPHFluid struct {
