+++ render/light/light.go
@@ -3,7 +3,7 @@ package light
 import (
 	"fmt"
 
-	"github.com/andewx/dieselfluid/math/mgl"
+	"github.com/andewx/dieselfluid/math/vector"
 )
 
 //----------------------------------------------------------------------------//
@@ -30,22 +30,22 @@ const (
 
 type Light interface {
 	Lx() *Source
-	Position() mgl.Vec
+	Position() vector.Vec
 	Type() int
-	SetRGB(mgl.Vec)
-	GetRGB() mgl.Vec
+	SetRGB(vector.Vec)
+	GetRGB() vector.Vec
 	GetFlux() float32
 	SetFlux(float32)
-	SetPos(mgl.Vec)
-	GetDir() mgl.Vec
-	SetDir(mgl.Vec)
+	SetPos(vector.Vec)
+	GetDir() vector.Vec
+	SetDir(vector.Vec)
 }
 
 //Source - Luminous Flux and Light Color in RGB Color Space. Systems dependent on
 //physical light interaction such as absorption, scattering should use light.SPD
 //Units are either based in Watts or Lumens per meter area
 type Source struct {
-	RGB  mgl.Vec
+	RGB  vector.Vec
 	Flux float32
 	Unit int
 }
@@ -54,7 +54,7 @@ type Source struct {
 //inverse squared attenuation factor given the total luminal output in lux.
 // Attenuated lights are synonymous with point lights.
 type Attenuated struct {
-	Pos mgl.Vec
+	Pos vector.Vec
 	Lum Source
 }
 
@@ -62,8 +62,8 @@ type Attenuated struct {
 //ray paths. Typically position is not a factor for directional lights but we
 //may model occlusions etc with a position
 type Directional struct {
-	Pos mgl.Vec
-	Dir mgl.Vec
+	Pos vector.Vec
+	Dir vector.Vec
 	Lum Source
 }
 
@@ -71,8 +71,8 @@ type Directional struct {
 //angle. For a point to recieve light it must be encapsulate by the projecting
 //solid angle. Lux is taken to a total output across this solid angle max
 type Area struct {
-	Pos     mgl.Vec
-	Norm    mgl.Vec
+	Pos     vector.Vec
+	Norm    vector.Vec
 	Cuttoff float32
 	Lum     Source
 }
@@ -83,7 +83,7 @@ func (p Attenuated) Lx() *Source {
 	return &p.Lum
 }
 
-func (p Attenuated) Position() mgl.Vec {
+func (p Attenuated) Position() vector.Vec {
 	return p.Pos
 }
 
@@ -91,12 +91,12 @@ func (p Attenuated) Type() int {
 	return ATTENUATED_LIGHT
 }
 
-func (p Attenuated) SetRGB(a mgl.Vec) {
+func (p Attenuated) SetRGB(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Lum.RGB = a
 	}
 }
-func (p Attenuated) GetRGB() mgl.Vec {
+func (p Attenuated) GetRGB() vector.Vec {
 	return p.Lum.RGB
 }
 func (p Attenuated) GetFlux() float32 {
@@ -105,16 +105,16 @@ func (p Attenuated) GetFlux() float32 {
 func (p Attenuated) SetFlux(a float32) {
 	p.Lum.Flux = a
 }
-func (p Attenuated) SetPos(a mgl.Vec) {
+func (p Attenuated) SetPos(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Pos = a
 	}
 }
-func (p Attenuated) GetDir() mgl.Vec {
+func (p Attenuated) GetDir() vector.Vec {
 	return nil
 }
 
-func (p Attenuated) SetDir(a mgl.Vec) {
+func (p Attenuated) SetDir(a vector.Vec) {
 
 }
 
@@ -124,7 +124,7 @@ func (p Directional) Lx() *Source {
 	return &p.Lum
 }
 
-func (p Directional) Position() mgl.Vec {
+func (p Directional) Position() vector.Vec {
 	return p.Pos
 }
 
@@ -132,12 +132,12 @@ func (p Directional) Type() int {
 	return DIRECTIONAL_LIGHT
 }
 
-func (p Directional) SetRGB(a mgl.Vec) {
+func (p Directional) SetRGB(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Lum.RGB = a
 	}
 }
-func (p Directional) GetRGB() mgl.Vec {
+func (p Directional) GetRGB() vector.Vec {
 	return p.Lum.RGB
 }
 func (p Directional) GetFlux() float32 {
@@ -146,16 +146,16 @@ func (p Directional) GetFlux() float32 {
 func (p Directional) SetFlux(a float32) {
 	p.Lum.Flux = a
 }
-func (p Directional) SetPos(a mgl.Vec) {
+func (p Directional) SetPos(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Pos = a
 	}
 }
-func (p Directional) GetDir() mgl.Vec {
+func (p Directional) GetDir() vector.Vec {
 	return p.Dir
 }
 
-func (p Directional) SetDir(a mgl.Vec) {
+func (p Directional) SetDir(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Dir = a
 	}
@@ -167,7 +167,7 @@ func (p Area) Lx() *Source {
 	return &p.Lum
 }
 
-func (p Area) Position() mgl.Vec {
+func (p Area) Position() vector.Vec {
 	return p.Pos
 }
 
@@ -175,12 +175,12 @@ func (p Area) Type() int {
 	return AREA_LIGHT
 }
 
-func (p Area) SetRGB(a mgl.Vec) {
+func (p Area) SetRGB(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Lum.RGB = a
 	}
 }
-func (p Area) GetRGB() mgl.Vec {
+func (p Area) GetRGB() vector.Vec {
 	return p.Lum.RGB
 }
 func (p Area) GetFlux() float32 {
@@ -189,15 +189,15 @@ func (p Area) GetFlux() float32 {
 func (p Area) SetFlux(a float32) {
 	p.Lum.Flux = a
 }
-func (p Area) SetPos(a mgl.Vec) {
+func (p Area) SetPos(a vector.Vec) {
 	if a != nil && len(a) == 3 {
 		p.Pos = a
 	}
 }
-func (p Area) GetDir() mgl.Vec {
+func (p Area) GetDir() vector.Vec {
 	return nil
 }
 
-func (p Area) SetDir(a mgl.Vec) {
+func (p Area) SetDir(a vector.Vec) {
 	fmt.Printf("Area light has no direction")
 }
