+++ render/light/probe.go
@@ -1,11 +1,11 @@
 package light
 
-import "github.com/andewx/dieselfluid/math/mgl"
+import "github.com/andewx/dieselfluid/math/vector"
 
 type Probe struct { //Use for light probes
 	//Lights   LightRig
 	Size     float32
 	Dims     [3]int
-	Position mgl.Vec
+	Position vector.Vec
 	Samples  int
 }
