+++ render/transform/transform.go
@@ -3,38 +3,39 @@ package transform
 import (
 	"math"
 
-	"github.com/andewx/dieselfluid/math/mgl"
+	"github.com/andewx/dieselfluid/math/matrix"
+	"github.com/andewx/dieselfluid/math/vector"
 )
 
 //Holds Transform Matrix
 type Transform struct {
-	Matrix mgl.Mat
+	Matrix matrix.Mat
 }
 
-func Apply(a mgl.Mat, b mgl.Mat) mgl.Mat {
+func Apply(a matrix.Mat, b matrix.Mat) matrix.Mat {
 	return a.MulM(b)
 }
 
 //Augments Transform Matrix with Translation
-func (T *Transform) Translate(vec mgl.Vec) mgl.Mat {
-	TransMat := mgl.Mat4(1.0)
+func (T *Transform) Translate(vec vector.Vec) matrix.Mat {
+	TransMat := matrix.Mat4(1.0)
 	TransMat[12] = vec[0]
 	TransMat[13] = vec[1]
 	TransMat[14] = vec[2]
-	T.Matrix = mgl.MulM(T.Matrix, TransMat)
+	T.Matrix = matrix.MulM(T.Matrix, TransMat)
 	return T.Matrix
 }
 
 //Applies tranform matrix to a vector
-func (T *Transform) Translate3VF(vec mgl.Vec) mgl.Vec {
+func (T *Transform) Translate3VF(vec vector.Vec) vector.Vec {
 	return T.Matrix.CrossVec(vec)
 }
 
 //Rotate applies matrix rotation of the angle theta which should be in radians,
 //The rotation axis is given by the Vector axis (l, m, n)
-func (T *Transform) Rotate(axis mgl.Vec, angle float32) mgl.Mat {
-	rot := mgl.Mat4(1.0)
-	axis = mgl.Norm(axis)
+func (T *Transform) Rotate(axis vector.Vec, angle float32) matrix.Mat {
+	rot := matrix.Mat4(1.0)
+	axis = vector.Norm(axis)
 	l := axis[0]
 	m := axis[1]
 	n := axis[2]
@@ -50,7 +51,7 @@ func (T *Transform) Rotate(axis mgl.Vec, angle float32) mgl.Mat {
 	rot[8] = l*n*mCos - m*sin
 	rot[9] = m*n*mCos + (l * sin)
 	rot[10] = n*n*mCos + cos
-	T.Matrix = mgl.MulM(T.Matrix, rot)
+	T.Matrix = matrix.MulM(T.Matrix, rot)
 	return T.Matrix
 }
 
@@ -67,7 +68,7 @@ func (T *Transform) EulerRotate(yaw float32, pitch float32, roll float32) {
 	cosX := float32(math.Cos(r))
 	sinX := float32(math.Sin(r))
 
-	euler := mgl.Mat{
+	euler := matrix.Mat{
 		cosZ * cosY, cosZ*sinY*sinX - sinZ*cosX, cosZ*sinY*cosX + sinZ*sinX, 0,
 		sinZ * cosY, sinZ*sinY*sinX + cosZ*cosX, sinZ*sinY*cosX - cosZ*sinX, 0,
 		-sinY, cosY * sinX, cosY * cosX, 0, 0, 0, 0, 1,
